refactor: 홈 화면 UI 및 데이터 모델 수정

+++ android/app/src/main/java/com/mulkkam/data/remote/model/response/MembersProgressInfoResponse.kt
@@ -24,7 +24,7 @@ fun MembersProgressInfoResponse.toDomain(): MembersProgressInfo =
     MembersProgressInfo(
         nickname = memberNickname,
         streak = streak,
-        achievementRate = achievementRate,
+        achievementRate = achievementRate.toFloat(),
         targetAmount = targetAmount,
         totalAmount = totalAmount,
         comment = comment,
